+++ iml-postgres/Cargo.toml
@@ -1,11 +1,13 @@
 [package]
-name = "iml-postgres"
-version = "0.3.0"
 authors = ["IML Team <iml@whamcloud.com>"]
 edition = "2018"
+name = "iml-postgres"
+version = "0.3.0"
 
 [dependencies]
-tokio-postgres = "0.5"
 futures = "0.3"
+iml-manager-env = {path = "../iml-manager-env", version = "0.3"}
+iml-wire-types = {path = "../iml-wire-types", version = "0.3"}
+sqlx = {git = "https://github.com/jgrund/sqlx", branch = "support-offline-workspaces", default-features = false, features = ["json", "macros", "offline", "postgres", "runtime-tokio", "time", "chrono"]}
+tokio-postgres = "0.5"
 tracing = "0.1"
-iml-manager-env = { path = "../iml-manager-env", version = "0.3" }
