+++ iml-services/iml-ntp/Cargo.toml
@@ -1,15 +1,15 @@
 [package]
-name = "iml-ntp"
-version = "0.3.0"
 authors = ["IML Team <iml@whamcloud.com>"]
 edition = "2018"
+name = "iml-ntp"
+version = "0.3.0"
 
 [dependencies]
 futures = "0.3"
-iml-orm = { path = "../../iml-orm", version = "0.3", features = ["postgres-interop"] }
-iml-rabbit = { path = "../../iml-rabbit", version = "0.3" }
-iml-service-queue = { path = "../iml-service-queue", version = "0.3" }
-iml-tracing = { version = "0.2", path="../../iml-tracing"}
-iml-wire-types = { path = "../../iml-wire-types", version = "0.3" }
-tokio = { version = "0.2", features = [ "rt-threaded", "blocking" ] }
+iml-postgres = {path = "../../iml-postgres", version = "0.3"}
+iml-rabbit = {path = "../../iml-rabbit", version = "0.3"}
+iml-service-queue = {path = "../iml-service-queue", version = "0.3"}
+iml-tracing = {version = "0.2", path = "../../iml-tracing"}
+iml-wire-types = {path = "../../iml-wire-types", version = "0.3"}
+tokio = {version = "0.2", features = ["rt-threaded", "blocking"]}
 tracing = "0.1"
