+++ iml-wire-types/Cargo.toml
@@ -8,6 +8,7 @@ version = "0.3.0"
 
 [dependencies]
 bytes = {version = "0.5", optional = true}
+chrono = {version = "0.4", features = ["wasmbind", "serde"]}
 im = {version = "14.3", features = ["serde"]}
 iml-api-utils = {path = "../iml-api-utils", version = "0.3"}
 iml-orm = {path = "../iml-orm", version = "0.3"}
